feat(duoplane): finish error handling for duoplane
Order submission button will now throw appropriate error if something
goes wrong. It should let the user know where in the process something
went wrong. I also extracted the logic into it's own hook.
